web consuming api 2023-0235

+++ SGHT.API/Controllers/RolUsuarioController.cs
@@ -3,6 +3,7 @@ using SGHT.API.Utils;
 using SGHT.Application.Dtos.RolUsuario;
 using SGHT.Application.Dtos.Usuarios;
 using SGHT.Application.Interfaces;
+using SGHT.Domain.Base;
 
 namespace SGHT.API.Controllers
 {
@@ -20,41 +21,80 @@ namespace SGHT.API.Controllers
         [HttpGet]
         public async Task<IActionResult> Get()
         {
-            var rolUsuarioList = await _rolUsuarioService.GetAll();
-            return HandleResponse(rolUsuarioList);
+            try
+            {
+                var rolUsuarioList = await _rolUsuarioService.GetAll();
+                return HandleResponse(rolUsuarioList);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
-            var rolUsuario = await _rolUsuarioService.GetById(id);
-            if (!rolUsuario.Success) return NotFound("user not found");
+            if (!IsIdValid(id)) return BadRequest("invalid id");
+            try
+            {
+                var rolUsuario = await _rolUsuarioService.GetById(id);
+                if (!rolUsuario.Success) return NotFound("user not found");
 
-            return HandleResponse(rolUsuario);
+                return HandleResponse(rolUsuario);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpPost("crear")]
         public async Task<IActionResult> Post(SaveRolUsuarioDto rolUsuario)
         {
-            var result = await _rolUsuarioService.Save(rolUsuario);
-
-            return HandleResponse(result);
+            if (string.IsNullOrWhiteSpace(rolUsuario.Descripcion)) return BadRequest("descripcion can't be null");
+            try
+            {
+                var result = await _rolUsuarioService.Save(rolUsuario);
+                return HandleResponse(result);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
 
         [HttpDelete("eliminar")]
         public async Task<IActionResult> Delete(DeleteRolUsuarioDto id)
         {
-            var result = await _rolUsuarioService.DeleteById(id);
+            if(!IsIdValid(id.IdRolUsuario)) return BadRequest("id is not valid");
+            try
+            {
+                var result = await _rolUsuarioService.DeleteById(id);
 
-            return HandleResponse(result);
+                return HandleResponse(result);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
 
         [HttpPatch("actualizar")]
         public async Task<IActionResult> Put(UpdateRolUsuarioDto rolUsuario)
         {
-            var result = await _rolUsuarioService.UpdateById(rolUsuario);
-            return HandleResponse(result);
+            if (!IsIdValid(rolUsuario.IdRolUsuario)) return BadRequest("id is not valid");
+            if (string.IsNullOrWhiteSpace(rolUsuario.Descripcion)) return BadRequest("descripcion can't be null");
+            try
+            {
+                var result = await _rolUsuarioService.UpdateById(rolUsuario);
+                return HandleResponse(result);
+            }
+            catch (Exception ex)
+            {
+                return HandleResponse(OperationResult.GetErrorResult("Internal Server Error", code: 500));
+            }
         }
     }
 }
